📦 move existing linter methods from helpers into linter/v1
