+++ lib/helpers.js
@@ -3,63 +3,15 @@
 import score from 'sb-string_score'
 import prettyPrintTypes from 'flow-language-server/lib/pkg/nuclide-flow-rpc/lib/prettyPrintTypes'
 import { Range, TextEditor, Point } from 'atom'
-import { prettyPrintError, mainMessageOfError } from './helpers-linter'
-import type { StatusObject, Location, CoverageObject, TypeAtPosObject } from './types'
+import type { CoverageObject, TypeAtPosObject } from './types'
+
+import { toLinterLocation, locationToRange } from './linter/v1'
 
 export const INIT_MESSAGE = 'flow server'
 export const RECHECKING_MESSAGE = 'flow is'
 
 export function getType(value: { value: string, type: string }) {
   return value.type && value.type.substr(0, 1) === '{' ? 'Object' : value.type || 'any'
-}
-
-export function locationToRange({ start, end }: Location) {
-  return new Range(
-    [start.line - 1, start.column - 1],
-    [end.line - 1, end.column],
-  )
-}
-export function toLinterLocation(loc: Location) {
-  return {
-    file: loc.source,
-    position: locationToRange(loc),
-  }
-}
-
-function toLinterReference(messages) {
-  for (let i = 1, length = messages.length; i < length; i++) {
-    const message = messages[i]
-    if (message.loc) {
-      return {
-        file: message.loc.source,
-        position: locationToRange(message.loc).start,
-      }
-    }
-  }
-  return null
-}
-
-export function toStatusLinterMessages(contents: string) {
-  const parsed: StatusObject = JSON.parse(contents)
-  if (!Array.isArray(parsed.errors) || !parsed.errors.length) {
-    return []
-  }
-
-  return parsed.errors.map((error) => {
-    const mainMsg = mainMessageOfError(error)
-    let excerpt = error.message.map(msg => msg.descr).join(' ')
-    if (error.operation && mainMsg === error.operation) {
-      excerpt = error.operation.descr + ' ' + excerpt
-    }
-
-    return {
-      severity: error.level === 'error' ? 'error' : 'warning',
-      location: mainMsg.loc && toLinterLocation(mainMsg.loc),
-      excerpt,
-      description: prettyPrintError(error),
-      reference: toLinterReference(error.message),
-    }
-  })
 }
 
 export function toCoverageLinterMessages(coverage: CoverageObject) {
