+++ lib/index.js
@@ -12,11 +12,11 @@ import {
   RECHECKING_MESSAGE,
   injectPosition,
   adjustPosition,
-  toStatusLinterMessages,
   toCoverageLinterMessages,
   toAutocompleteSuggestions,
   toDatatip,
 } from './helpers'
+import * as LinterV1 from './linter/v1'
 import CoverageView from './coverage-view'
 import { toOutline } from './outline'
 import type { OutlineOptions } from './outline/types'
@@ -153,7 +153,7 @@ export default {
           }
         }
 
-        return toStatusLinterMessages(result)
+        return LinterV1.toLinterMessages(result)
       },
     }
     return linter
