+++ lib/linter/v1/index.js
@@ -1,0 +1,56 @@
+/* @flow */
+
+import { Range, TextEditor, Point } from 'atom'
+import { mainMessageOfError, prettyPrintError } from './pretty'
+
+import type { StatusMessage, StatusObject } from './types'
+import type { Location } from '../../types'
+
+export function locationToRange({ start, end }: Location) {
+  return new Range(
+    [start.line - 1, start.column - 1],
+    [end.line - 1, end.column],
+  )
+}
+export function toLinterLocation(loc: Location) {
+  return {
+    file: loc.source,
+    position: locationToRange(loc),
+  }
+}
+
+function toLinterReference(messages: StatusMessage[]) {
+  for (let i = 1, length = messages.length; i < length; i++) {
+    const message = messages[i]
+    if (message.loc) {
+      return {
+        file: message.loc.source,
+        position: locationToRange(message.loc).start,
+      }
+    }
+  }
+  return null
+}
+
+export function toLinterMessages(contents: string) {
+  const parsed: StatusObject = JSON.parse(contents)
+  if (!Array.isArray(parsed.errors) || !parsed.errors.length) {
+    return []
+  }
+
+  return parsed.errors.map((error) => {
+    const mainMsg = mainMessageOfError(error)
+    let excerpt = error.message.map(msg => msg.descr).join(' ')
+    if (error.operation && mainMsg === error.operation) {
+      excerpt = error.operation.descr + ' ' + excerpt
+    }
+
+    return {
+      severity: error.level === 'error' ? 'error' : 'warning',
+      location: mainMsg.loc && toLinterLocation(mainMsg.loc),
+      excerpt,
+      description: prettyPrintError(error),
+      reference: toLinterReference(error.message),
+    }
+  })
+}
