+++ lib/linter/v1/pretty.js
@@ -5,7 +5,8 @@
 // and adjusted to output nicely formatted markdown.
 
 import { format } from 'util'
-import type { StatusError, StatusExtra, StatusMessage, Location } from './types'
+import type { StatusError, StatusExtra, StatusMessage } from './types'
+import type { Location } from '../../types'
 
 const regexMarkdownChars = /[*#()[\]<>_]/g
 const escapeChars = { '<': '&lt;', '>': '&gt;' }
